+++ srcs/dist.c
@@ -12,7 +12,7 @@
 
 #include "cub3d.h"
 
-float	dist(float x1, float y1, float x2, float y2)
+double	dist(double x1, double y1, double x2, double y2)
 {
 	return (sqrt((x2 - x1) * (x2 - x1) + (y2 - y1) * (y2 - y1)));
 }
